moved some code around added comments and renamed some variables to make the code easier to understand migrated to new image_util build system improved tests updated copyright headers to 2018 updated webhook version
