+++ test/images/webhook/VERSION
@@ -1,1 +1,1 @@
-1.10v2
+1.12v1
