+++ test/images/webhook/patch_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2017 The Kubernetes Authors.
+Copyright 2018 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -26,37 +26,87 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
-func TestJSONPatchForConfigMap(t *testing.T) {
-	cm := corev1.ConfigMap{
-		Data: map[string]string{
-			"mutation-start": "yes",
+func TestPatches(t *testing.T) {
+	testCases := []struct {
+		patch    string
+		initial  interface{}
+		expected interface{}
+		toTest   interface{}
+	}{
+		{
+			patch: configMapPatch1,
+			initial: corev1.ConfigMap{
+				Data: map[string]string{
+					"mutation-start": "yes",
+				},
+			},
+			expected: &corev1.ConfigMap{
+				Data: map[string]string{
+					"mutation-start":   "yes",
+					"mutation-stage-1": "yes",
+				},
+			},
+		},
+		{
+			patch: configMapPatch2,
+			initial: corev1.ConfigMap{
+				Data: map[string]string{
+					"mutation-start": "yes",
+				},
+			},
+			expected: &corev1.ConfigMap{
+				Data: map[string]string{
+					"mutation-start":   "yes",
+					"mutation-stage-2": "yes",
+				},
+			},
+		},
+
+		{
+			patch: podsInitContainerPatch,
+			initial: corev1.Pod{
+				Spec: corev1.PodSpec{
+					InitContainers: []corev1.Container{},
+				},
+			},
+			expected: &corev1.Pod{
+				Spec: corev1.PodSpec{
+					InitContainers: []corev1.Container{
+						{
+							Image:     "webhook-added-image",
+							Name:      "webhook-added-init-container",
+							Resources: corev1.ResourceRequirements{},
+						},
+					},
+				},
+			},
 		},
 	}
-	cmJS, err := json.Marshal(cm)
-	if err != nil {
-		t.Fatal(err)
+	for _, testcase := range testCases {
+		objJS, err := json.Marshal(testcase.initial)
+		if err != nil {
+			t.Fatal(err)
+		}
+		patchObj, err := jsonpatch.DecodePatch([]byte(testcase.patch))
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		patchedJS, err := patchObj.Apply(objJS)
+		if err != nil {
+			t.Fatal(err)
+		}
+		objType := reflect.TypeOf(testcase.initial)
+		objTest := reflect.New(objType).Interface()
+		err = json.Unmarshal(patchedJS, objTest)
+		if err != nil {
+			t.Fatal(err)
+		}
+		if !reflect.DeepEqual(objTest, testcase.expected) {
+			t.Errorf("\nexpected %#v\n, got %#v", testcase.expected, objTest)
+		}
 	}
 
-	patchObj, err := jsonpatch.DecodePatch([]byte(patch1))
-	if err != nil {
-		t.Fatal(err)
-	}
-	patchedJS, err := patchObj.Apply(cmJS)
-	patchedObj := corev1.ConfigMap{}
-	err = json.Unmarshal(patchedJS, &patchedObj)
-	if err != nil {
-		t.Fatal(err)
-	}
-	expected := corev1.ConfigMap{
-		Data: map[string]string{
-			"mutation-start":   "yes",
-			"mutation-stage-1": "yes",
-		},
-	}
-
-	if !reflect.DeepEqual(patchedObj, expected) {
-		t.Errorf("\nexpected %#v\n, got %#v", expected, patchedObj)
-	}
 }
 
 func TestJSONPatchForUnstructured(t *testing.T) {
@@ -74,7 +124,7 @@ func TestJSONPatchForUnstructured(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	patchObj, err := jsonpatch.DecodePatch([]byte(patch1))
+	patchObj, err := jsonpatch.DecodePatch([]byte(configMapPatch1))
 	if err != nil {
 		t.Fatal(err)
 	}
